+++ tests/validation-tests/api/mod.rs
@@ -1,4 +1,5 @@
 mod binding_arrays;
 mod buffer;
 mod buffer_slice;
+mod instance;
 mod texture;